Bugfix: Use sint64 instead of sint32 for data and index sizes in advanced properties dialog.  Should fix issue #1738943.  I couldn't find any Db.TField.AsInt64 function, so values for a single file will still wrap if the file is larger than 4GB.  Could be fixed by selecting from information_schema and having the database server do the division by 1024..

+++ tbl_properties.pas
@@ -52,8 +52,8 @@ var
   ts: TTabSheet;
   list: TSortListView;
   ListTables: TSortListView;
-  datasize: Integer;
-  indexsize: Integer;
+  datasize: Int64;
+  indexsize: Int64;
   isSelected: Boolean;
   FieldList: TStringList;
   splitter: TSplitter;
